requirements.txt
absl-py
accelerate
annotated-doc
anyio
av
certifi
charset-normalizer
click
editdistance
filelock
fsspec
h11
hf-xet
httpcore
httpx
huggingface_hub
idna
Jinja2
joblib
markdown-it-py
MarkupSafe
mdurl
mpmath
networkx
nltk
numpy
packaging
peft
pillow
psutil
Pygments
PyMuPDF
PyYAML
qwen-vl-utils
regex
requests
rich
rouge_score
safetensors
setuptools
shellingham
six
sympy
tokenizers
torch
torchvision
tqdm
transformers
typer
typing_extensions
urllib3
